api/requirements.txt
Django==5.2.1
gunicorn==23.0.0
psycopg2==2.9.10
djangorestframework==3.16.0
celery==5.5.2
redis==6.2.0
Faker==37.3.0
